refactor attachable entity
