+++ client/src/pages/Types/coreEntity/attachableEntity.ts
@@ -1,10 +1,14 @@
 import { assertUnreachable } from '../../Component/Visualizer/util/util';
 import { EntityType } from '../entity/baseEntity';
 import { NodeEntity } from '../entity/nodeEntity';
+import { EntityConcrete } from '../frontendType';
 import { Circle, ShapeType } from '../geometry/geometry';
 
+/**
+ * Attachable Entity refers to the type of entity that has a shape without infers from other entity.
+ * An example is NodeEntity
+ */
 export type AttachableEntity = NodeEntity;
-
 export function getAttachableEntityShape(entity: AttachableEntity): Circle {
   switch (entity.type) {
     case EntityType.NODE:
@@ -21,4 +25,17 @@ export function getAttachableEntityShape(entity: AttachableEntity): Circle {
       assertUnreachable(entity.type);
   }
   throw new Error('Unreachable');
+}
+
+export function isAttachableEntity(entity: EntityConcrete): entity is AttachableEntity {
+  switch (entity.type) {
+    case EntityType.NODE:
+      return true;
+    case EntityType.EDGE:
+    case EntityType.POINTER:
+      return false;
+    default:
+      assertUnreachable(entity);
+      return false;
+  }
 }
