--- client/src/pages/Types/coreEntity/concreteEntity.ts
@@ -1,17 +1,0 @@
-import { assertUnreachable } from '../../Component/Visualizer/util/util';
-import { EntityType } from '../entity/baseEntity';
-import { NodeEntity } from '../entity/nodeEntity';
-import { EntityConcrete } from '../frontendType';
-import { AttachableEntity } from './attachableEntity';
-
-export function isAttachableEntity(entity: EntityConcrete): entity is AttachableEntity {
-  switch (entity.type) {
-    case EntityType.NODE:
-      return true;
-    case EntityType.EDGE:
-    case EntityType.POINTER:
-      return false;
-    default:
-      assertUnreachable(entity);
-  }
-}
